Fixed statically accessed keyboard inputs. Added FPS counter. Added new FPS Counter Flag. Placed debug logs inside of checks for debugMode flag.

+++ src/com/vex/Scene.java
@@ -7,6 +7,6 @@ public abstract class Scene
 
     }   
     
-    public abstract void update(float deltaTime);
+    public abstract void update(double deltaTime);
 
 }
